// apps/frontend/src/api/generated/users/model/shortItem.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * документация для api для mia coffee bot
 * Описание api для демонстрации документации в формате JSON
 * OpenAPI spec version: 1.0.0
 */

export interface ShortItem {
  favorite: boolean;
  id: number;
  /** @nullable */
  image: string | null;
  title: string;
}